+++ testdata/app/presenters/users/work/overall_presenter.rb
@@ -1,0 +1,11 @@
+module Users
+  module Work
+    class OverallPresenter
+      attr_reader :worker
+
+      def total_hours
+        worker.shifts.sum(&:hours)
+      end
+    end
+  end
+end